fix(api-page-builder-import-export): improve preview image handling
