+++ packages/api-page-builder-import-export/src/export/process/exporters/FormExporter.ts
@@ -1,0 +1,39 @@
+import Zipper from "~/export/zipper";
+import { FbForm } from "@webiny/api-form-builder/types";
+import { File } from "@webiny/api-file-manager/types";
+
+export interface ExportedFormData {
+    form: Pick<
+        FbForm,
+        "name" | "status" | "version" | "fields" | "layout" | "settings" | "triggers"
+    >;
+    files: File[];
+}
+
+export class FormExporter {
+    async execute(form: FbForm, exportFormsDataKey: string) {
+        const formData = {
+            form: {
+                name: form.name,
+                status: form.status,
+                version: form.version,
+                fields: form.fields,
+                layout: form.layout,
+                settings: form.settings,
+                triggers: form.triggers
+            }
+        };
+        const formDataBuffer = Buffer.from(JSON.stringify(formData));
+
+        const zipper = new Zipper({
+            exportInfo: {
+                files: [],
+                name: form.name,
+                dataBuffer: formDataBuffer
+            },
+            archiveFileKey: exportFormsDataKey
+        });
+
+        return zipper.process();
+    }
+}
